:truck: refactor: Response DTO 패키지 이동
