+++ backend/src/main/java/cotato/backend/domains/post/service/PostService.java
@@ -3,7 +3,7 @@ package cotato.backend.domains.post.service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import cotato.backend.domains.post.dto.PostResponse;
+import cotato.backend.domains.post.dto.response.PostResponse;
 import cotato.backend.domains.post.dto.request.SavePostsByExcelRequest;
 import cotato.backend.domains.post.dto.request.SaveSinglePostRequest;
